remove BoxB::ExtractRegisterAs in favor of the separate ExtractRegisterAs;
move RegisterId to ergo_box::register module;
