+++ ergo-lib/src/ast.rs
@@ -1,8 +1,8 @@
 //! AST for ErgoTree
 
-pub(crate) mod box_methods;
 pub(crate) mod coll_methods;
 pub(crate) mod expr;
+pub(crate) mod extract_reg_as;
 pub(crate) mod global_vars;
 pub(crate) mod method_call;
 pub(crate) mod ops;
