+++ ergo-lib/src/ast/expr.rs
@@ -3,10 +3,10 @@ use core::fmt;
 use crate::serialization::op_code::OpCode;
 use crate::types::stype::SType;
 
-use super::box_methods::BoxM;
 use super::coll_methods::CollM;
 use super::constant::Constant;
 use super::constant::ConstantPlaceholder;
+use super::extract_reg_as::ExtractRegisterAs;
 use super::global_vars::GlobalVars;
 use super::method_call::MethodCall;
 use super::ops;
@@ -28,8 +28,6 @@ pub enum Expr {
     PredefFunc(PredefFunc),
     /// Collection type methods
     CollM(CollM),
-    /// Box methods
-    BoxM(BoxM),
     Context,
     // Global(Global),
     /// Predefined global variables
@@ -42,6 +40,8 @@ pub enum Expr {
     BinOp(ops::BinOp, Box<Expr>, Box<Expr>),
     /// Option get method
     OptionGet(Box<OptionGet>),
+    /// Extract register's value (box.RX properties)
+    ExtractRegisterAs(Box<ExtractRegisterAs>),
 }
 
 impl Expr {
